distage-framework: working JS roleapp, further improvements should be done separately, closes #1290
