--- distage/distage-framework/.js/src/test/scala/izumi/distage/framework/SJSFrameworkTest.scala
@@ -1,11 +1,0 @@
-package izumi.distage.framework
-
-import org.scalatest.wordspec.AnyWordSpec
-
-class SJSFrameworkTest extends AnyWordSpec {
-  "distage-framework" should {
-    "run under SJS" in {
-
-    }
-  }
-}
